+++ src/app/admin/admin-products/admin-products.component.css
@@ -1,0 +1,35 @@
+.modalContainer{
+    display: block;
+    align-items: center;
+    border: 1px solid black;
+    position:absolute;
+    z-index:3;
+    background-color: rgba(0, 0, 0, .5);
+    width: 100%;
+    height: 100%;
+  }
+  /* input{
+    width:10rem;
+    background-color: aquamarine;
+  } */
+
+  input, button {
+    width : 150px;
+    padding: 10px;
+    margin: 0;
+    box-sizing: border-box;
+  }
+  .modalContent{
+
+    background-color:#EFB45D;
+    height: 30rem;
+    width: 30rem;
+    margin: 100px auto;
+    
+  }
+	.modalContent label{
+		
+		margin-top: 5rem;
+    
+  }
+
