+++ src/app/admin/admin.component.css
@@ -1,15 +1,79 @@
-.modalContainer{
-  display: block;
-  align-items: center;
-  border: 1px solid black;
-  position:absolute;
-  z-index:3;
-}
-input{
-  width:10rem;
-  background-color: aquamarine;
-}
-
 /* .modalContent{
   width: ;;
 } */
+/* .mainContainer{
+  height: 80vh;
+  flex-wrap: wrap;
+} */
+.btnContainer{
+  /* display: flex; */
+  justify-content: center;
+  /* align-items: start; */
+  /* position:absolute; */
+  top: 55px;
+  /* margin: 0 auto; */
+  /* margin-bottom: 10vh; */
+  
+}
+
+.card{
+  display: flex;
+  justify-content: center;
+  text-align: center;
+  border-radius: 20px;
+  background: white;
+  position: relative;
+  padding: 1rem;
+  border: 2px solid #c3c6ce;
+  transition: 0.5s ease-out;
+  overflow: visible;
+  height: 18rem;
+  width: 50%;
+  margin: 10px;
+}
+
+.card-button {
+  transform: translate(-50%, 1%);
+  width: 60%;
+  border-radius: 1rem;
+  border: none;
+  background-color:  #730202;
+  color: #fff;
+  font-size: 1rem;
+  padding: .5rem 1rem;
+  position: absolute;
+  opacity: 1;
+  transition: 0.3s ease-out;
+ }
+
+ .card:hover {
+  border-color: #730202;
+  box-shadow: 0 4px 18px 0 rgba(0, 0, 0, 0.25);
+ }
+
+ .btnFood {
+  /* position: relative;
+  display: inline-block; */
+  margin: 5px;
+  padding: 10px;
+  text-align: center;
+  font-size: 18px;
+  letter-spacing: 1px;
+  text-decoration: none;
+  color: black;
+  background: transparent;
+  cursor: pointer;
+  /* transition: ease-out 0.5s; */
+  border: 3px solid black;
+  border-radius: 10px;
+  box-shadow: inset 0 0 0 0 orange;
+}
+
+.btnFood:hover {
+  color: white;
+  box-shadow: inset 0 -100px 0 0 black;
+}
+
+.btnFood:active {
+  transform: scale(0.9);
+}
