+++ src/app/nav/nav.component.html
@@ -1,4 +1,4 @@
-<router-outlet></router-outlet>
+<!-- <router-outlet></router-outlet> -->
 <nav class="navbar navbar-expand-lg bg-body-tertiary">
 	<div class="container-fluid">
 		<a class="navbar-brand fw-bold" href="#"> <fa-icon [icon]="faBurger"></fa-icon>
